damage

+++ src/characters/functions/dealDamage.ts
@@ -1,5 +1,5 @@
 import { GameState } from '../../game/GameState'
-import { DamageTypes } from '../../items/Item'
+import { DamageData, DamageTypes } from '../../items/Item'
 import { CharacterAdapter } from '../characterAdapter'
 import { selectCharacterArmour } from '../selectors/armourSelector'
 import { getDamageMulti } from './getDamageMulti'
@@ -8,22 +8,27 @@ import { kill } from './kill'
 export function dealDamage(
     state: GameState,
     targetId: string,
-    damage: number,
-    damageType: DamageTypes
+    damageData: DamageData
 ): { state: GameState; killed: boolean } {
-    const target = CharacterAdapter.selectEx(state.characters, targetId)
-    const armour = selectCharacterArmour(targetId, damageType)(state)
     let killed = false
-    const multi = getDamageMulti(damage, armour)
-    const damageTaken = damage * multi
+    const target = CharacterAdapter.selectEx(state.characters, targetId)
+    Object.entries(damageData).forEach((kv) => {
+        if (killed) return
 
-    const health = Math.max(0, Math.floor(target.health - damageTaken))
-    if (health < 0.0001) {
-        state = kill(state, targetId)
-        killed = true
-    } else {
-        state = { ...state, characters: CharacterAdapter.update(state.characters, targetId, { health }) }
-    }
+        const damageType: DamageTypes = kv[0] as DamageTypes
+        const damage = kv[1]
+        const armour = selectCharacterArmour(targetId, damageType)(state)
 
+        const multi = getDamageMulti(damage, armour)
+        const damageTaken = damage * multi
+
+        const health = Math.max(0, Math.floor(target.health - damageTaken))
+        if (health < 0.0001) {
+            state = kill(state, targetId)
+            killed = true
+        } else {
+            state = { ...state, characters: CharacterAdapter.update(state.characters, targetId, { health }) }
+        }
+    })
     return { state, killed }
 }
